Included in this GitHub file is our classes developed for our Final Project starting with our Dispenser Class which manages the vending Machine and the Product class which manages the inventory within the vending machine.
We then created 2 abstract classes of the Product class for Drinks and Snacks. We took the Snacks class further as an abstract class allowing for subclasses to be developed for Chips, Candy, and Gum.

In this milestone, we implemented the Comparable interface to the drinks and snacks classes allowing for sorting of the products in the inventory of each class. Using the CompareTo() method allows the products to be sorted following alphabetical ordering rules.

Our main class was updated to create 2 arraylists for each class and implemented the sort. To run the program, just click Run.